requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
python-dotenv==1.1.0
pydantic==2.11.2
email-validator==2.1.1  # Required for Pydantic's EmailStr validator
requests==2.31.0  # Required for OpenAI service HTTP requests
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
openai==1.71.0
PyPDF2==3.0.1
python-docx==1.0.1
Jinja2==3.1.3
python-multipart==0.0.7
numpy==1.26.4  # Compatible with pandas 2.2.2
pandas==2.2.2
openpyxl==3.1.2
pyarrow==15.0.2
boto3==1.37.8  # AWS S3 SDK
psycopg2-binary==2.9.9  # PostgreSQL adapter